requirements.txt
steamlit
pandas
openai
python-dotenv